refactor(step1): menus display 판단 로직 개선
